Budget Planning Example - Clinical Trial
This is a basic budget designed to highlight core costs of a clinical trial.  Sub-categories are required to break down the actual costs in a more complex budget that would be used in a protocol, but this should be a good starting point to understand what is expected in the budget of a clinical trial.